// apps/web/src/app/api/models/immobilie-antwort-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ImmobilieAntwortDto {
  beschreibung: string;
  hausnummer: string;
  id: string;
  land: 'Deutschland' | 'Italien' | 'Frankreich';
  name: string;
  postleitzahl: string;
  stadt: string;
  strasse: string;
}